Organize helper functions to provide a cleaner way of setting the visibility
