+++ common/navigation/src/main/res/menu/bottom_nav_menu.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
-    <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black"
-        android:title="@string/title_home" />
+<!--    <item-->
+<!--        android:id="@+id/navigation_home"-->
+<!--        android:icon="@drawable/ic_home_black"-->
+<!--        android:title="@string/title_home" />-->
 
     <item
         android:id="@+id/navigation_dashboard"
